correcao funcionalidade ditribuir e e recolher

+++ src/test/java/supporte/Distribuir.java
@@ -17,6 +17,7 @@ public class Distribuir {
     private static File pastacotacao;
     private static File pastaimage;
     private static PrintWriter gravarArq;
+    private boolean existenrcotacao;
 
     public void distribuirArquivos() throws IOException {
         File destpastamae = new File("/home/robertinho/suhai_perfil");
@@ -31,13 +32,15 @@ public class Distribuir {
         pastaimage = CriarPastas.criarPastaimage();
 
 
-
-        //diretorio de onde vai copiar a pasta mae
-        File sourcefreenaspastamae = new File("/run/user/1000/gvfs/smb-share:server=freenas.local,share=arquivosgerais/Macro certa/selenium/outros/ferramentas");
-        //diretorio de onde vai copiar a pasta na vm
-        destpastamae = new File("/home/robertinho/ferramentas");
-        destpastamae.mkdir();
-        FileUtils.copyDirectory(sourcefreenaspastamae, destpastamae);
+        existenrcotacao = GerenciaNrCotacao.VerificaCotacao();
+        if (existenrcotacao == false) {
+            //diretorio de onde vai copiar a pasta mae
+            File sourcefreenaspastamae = new File("/run/user/1000/gvfs/smb-share:server=freenas.local,share=arquivosgerais/E/LABSEG/apolice/result/SUHAI/PRONTOS/mover");
+            //diretorio de onde vai copiar a pasta na vm
+            destpastamae = new File("/home/robertinho/ferramentas/mover");
+            destpastamae.mkdir();
+            FileUtils.copyDirectory(sourcefreenaspastamae, destpastamae);
+        }
 
         //cria arquivo com o nome da vm caso nao exista
         gravarArq = PegarNomeVm.criarNomeVM();
@@ -52,13 +55,16 @@ public class Distribuir {
             FileUtils.deleteDirectory(diretorio);
         }
 
-        File directoryresumo = new File("/run/user/1000/gvfs/smb-share:server=freenas.local,share=arquivosgerais/E/LABSEG/apolice/result/SUHAI/PRONTOS/cotacao/");
-        File destresumo = new File("/home/robertinho/suhai_perfil/cotacao/");
-        destresumo.mkdir();
-        FileUtils.copyDirectory(directoryresumo,destresumo);
+        if (existenrcotacao == false) {
+            File directoryresumo = new File("/run/user/1000/gvfs/smb-share:server=freenas.local,share=arquivosgerais/E/LABSEG/apolice/result/SUHAI/PRONTOS/cotacao/");
+            File destresumo = new File("/home/robertinho/suhai_perfil/cotacao/");
+            destresumo.mkdir();
+            FileUtils.copyDirectory(directoryresumo, destresumo);
+        }
+
         File directoryimage = new File("/run/user/1000/gvfs/smb-share:server=freenas.local,share=arquivosgerais/E/LABSEG/apolice/result/suhai/PRONTOS/images/");
         File destimage = new File("/home/robertinho/suhai_perfil/images/");
         destimage.mkdir();
-        FileUtils.copyDirectory(directoryimage,destimage);
+        FileUtils.copyDirectory(directoryimage, destimage);
     }
 }
